time_and_memory/runit.c

+++ time_and_memory/u_template.h
@@ -11,6 +11,6 @@ typedef struct
 } triangle_t;
 
 struct sieve_t {
-    int long long n;
+    int n;
     char *s;
 };
